torchlake/graph/requirements.txt
torch_geometric
networkx